Move TrackTooltip to it's own component
